% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet_to_csv.R
\name{add_sheet_to_csv}
\alias{add_sheet_to_csv}
\title{Combine a new dataframe as a new sheet to an existing csv into a Excel file}
\usage{
add_sheet_to_csv(file, new_sheet_name, dataframe, commit = TRUE)
}
\arguments{
\item{file}{path to csv file}

\item{new_sheet_name}{name of new sheet}

\item{dataframe}{dataframe tp write to Excel file}

\item{commit}{If TRUE, performs a git commit before and after the new sheet is added}
}
\value{
Returns the name of the new xlsx filename
}
\description{
Combine a new dataframe as a new sheet to an existing csv into a Excel file
}
