% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet_to_excel.R
\name{add_sheet_to_excel}
\alias{add_sheet_to_excel}
\title{Add a dataframe to an existing Excel file}
\usage{
add_sheet_to_excel(file, new_sheet_name, dataframe, commit = TRUE)
}
\arguments{
\item{file}{path to Excel file}

\item{new_sheet_name}{name of new sheet}

\item{dataframe}{dataframe tp write to Excel file}

\item{commit}{If TRUE, performs a git commit before and after the new sheet is added}
}
\description{
Add a dataframe to an existing Excel file
}
