% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_csv.R
\name{append_csv}
\alias{append_csv}
\title{Append a dataframe to a csv by binding rows to prevent column mismatch}
\usage{
append_csv(csv_fn, dataframe, log = TRUE, log_details = "")
}
\arguments{
\item{fn}{filename to be appended}
}
\description{
Append a dataframe to a csv by binding rows to prevent column mismatch
}
