% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-pkg.R
\name{args}
\alias{args}
\title{Broca Arguments}
\arguments{
\item{data}{A dataframe or tibble.}

\item{log}{(Deprecated) Log this function call to a log.md file?}

\item{log_details}{(Deprecated) If log is TRUE, additional notes to include with the log entry.}

\item{header}{True if data has a header.}

\item{quote}{Single character used to quote strings.}

\item{file}{File path.}
}
\description{
Broca Arguments
}
