% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_to_rmd.R
\name{convert_docx_to_rmd}
\alias{convert_docx_to_rmd}
\title{Converts all docx files in the working directory to markdown}
\usage{
convert_docx_to_rmd(input_docx)
}
\arguments{
\item{input_docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
