% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xlsx_to_csv.R
\name{convert_xlsx_to_csv}
\alias{convert_xlsx_to_csv}
\title{Convert a single sheet Excel file to a csv}
\usage{
convert_xlsx_to_csv(xlsx_file, remove = FALSE)
}
\arguments{
\item{xlsx_file}{path to Excel file}

\item{remove}{TRUE if the source Excel file should be deleted. Defaults to FALSE.}
}
\description{
This function only operates on Excel files with only 1 tab
}
