% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{copy_cb}
\alias{copy_cb}
\title{Copy a Dataframe to Clipboard}
\usage{
copy_cb(data, log = "DEPRECATED", log_details = "")
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{log}{(Deprecated) Log this function call to a log.md file?}

\item{log_details}{(Deprecated) If log is TRUE, additional notes to include with the log entry.}
}
\description{
Copy a Dataframe to Clipboard
}
\examples{
library(tidyverse)

# Test Data
test_data <-
        tibble(A = 1:3,
               B = c("Apples", "Bananas", "Oranges"),
               C = factor(c("Apples", "Bananas", "Oranges")),
               D = runif(3))


# Copy to Clipboard
copy_cb(data = test_data)

# Read Clipboard
read_cb()

# Check Datatypes
sapply(test_data, FUN = class)
sapply(read_cb(), FUN = class) # `read_cb()` function mutates all factors to character before returning

# Read Clipboard Shortcut for Headerless Data
read_cb_hl()

# Read Clipboard as a Tribble for Copy and Paste
read_cb_trbl()
read_cb_trbl(header = FALSE)
read_cb_trbl(quote = "'")
read_cb_trbl(header = FALSE, quote = "'")
}
\seealso{
Other clipboard functions: 
\code{\link{copy_to_clipboard}()},
\code{\link{read_cb_hl}()},
\code{\link{read_cb_trbl}()},
\code{\link{read_cb}()},
\code{\link{read_clipboard}()}
}
\concept{clipboard functions}
