% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_to_gfm.R
\name{docx_to_gfm}
\alias{docx_to_gfm}
\title{Converts all docx files in the working directory to markdown}
\usage{
docx_to_gfm(docx, log = TRUE, log_details = "")
}
\arguments{
\item{docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
