% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_to_md.R
\name{docx_to_md}
\alias{docx_to_md}
\title{Converts all docx files in the working directory to markdown}
\usage{
docx_to_md(input_docx, log = TRUE, log_details = "")
}
\arguments{
\item{docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
