% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{docx_to_rmd}
\alias{docx_to_rmd}
\title{Converts all docx files in the working directory to markdown}
\usage{
docx_to_rmd(input_docx, log = TRUE, log_details = "")
}
\arguments{
\item{input_docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
