% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_csv.R
\name{excel_to_csv}
\alias{excel_to_csv}
\title{Converts all xlsx files in the working directory to csv}
\usage{
excel_to_csv(input_xlsx)
}
\arguments{
\item{input_xlsx}{name of the input xlsx filename}
}
\description{
Converts all xlsx files in the working directory to csv
}
