% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{is_csv}
\alias{is_csv}
\title{Does the file have a csv extension?}
\usage{
is_csv(file)
}
\arguments{
\item{file}{File path.}
}
\description{
Check if the file has a name that matches the ".csv"
file extension pattern.
}
\seealso{
Other logical functions: 
\code{\link{is_excel}()}

Other csv functions: 
\code{\link{read_csv_as_rmd_table}()},
\code{\link{simply_read_csv}()},
\code{\link{simply_write_csv}()},
\code{\link{write_temp_csv}()}
}
\concept{csv functions}
\concept{logical functions}
