% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{is_excel}
\alias{is_excel}
\title{Does the file have an xlsx extension?}
\usage{
is_excel(file)
}
\arguments{
\item{file}{File path.}
}
\description{
Check if the file has a name that matches the ".xlsx" file extension pattern.
}
\seealso{
Other logical functions: 
\code{\link{is_csv}()}

Other excel functions: 
\code{\link{read_full_excel}()},
\code{\link{write_full_excel}()},
\code{\link{write_temp_xlsx}()}
}
\concept{excel functions}
\concept{logical functions}
