% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{knit_asis}
\alias{knit_asis}
\title{FUNCTION_TITLE}
\usage{
knit_asis(table)
}
\arguments{
\item{table}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
