% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{knit_asis_rmd_table_value}
\alias{knit_asis_rmd_table_value}
\title{FUNCTION_TITLE}
\usage{
knit_asis_rmd_table_value(
  heading_level = 2,
  csv_rmd_table,
  options = NULL,
  envir = parent.frame(),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{heading_level}{PARAM_DESCRIPTION, Default: 2}

\item{options}{PARAM_DESCRIPTION, Default: NULL}

\item{envir}{PARAM_DESCRIPTION, Default: parent.frame()}

\item{quiet}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[knitr]{opts_knit}},\code{\link[knitr]{knit_child}}
 \code{\link[glue]{glue}}
}
