% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_this.R
\name{log_this}
\alias{log_this}
\title{Universal log function to use to log items}
\usage{
log_this(path_to_file, activity_type, function_used = "", details = "")
}
\description{
Universal log function to use to log items
}
