% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_trbl}
\alias{make_trbl}
\title{Make a Tribble Skeleton}
\usage{
make_trbl(data, quote = "\\"")
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{quote}{Single character used to quote strings.}
}
\description{
Make a Tribble Skeleton
}
\seealso{
\code{\link[dplyr]{mutate_if}}
}
\concept{make functions}
