% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeTribble}
\alias{makeTribble}
\title{Make a tribble skeleton}
\usage{
makeTribble(.data)
}
\arguments{
\item{.data}{data frame}
}
\description{
Make a tribble skeleton
}
\seealso{
\code{\link[dplyr]{mutate_all}}
}
