% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_extension.R
\name{makeXL}
\alias{makeXL}
\title{FUNCTION_TITLE}
\usage{
makeXL()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xlsx]{Workbook}}
}
