% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr_file.R
\name{ocr_file}
\alias{ocr_file}
\title{Tesseract's ocr function, but the png outputs are deleted after execution}
\usage{
ocr_file(pdf_fn, pages = NULL, txt_fn)
}
\arguments{
\item{pdf_fn}{full path to the input pdf name}

\item{pages}{pages to ocr}

\item{txt_fn}{full path to output txt file}
}
\description{
Tesseract's ocr function, but the png outputs are deleted after execution
}
