% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xlsx_to_csvs.R
\name{parse_xlsx_to_csvs}
\alias{parse_xlsx_to_csvs}
\title{Parses tabs in an Excel to csvs}
\usage{
parse_xlsx_to_csvs(xlsx_fn, remove = FALSE)
}
\arguments{
\item{xlsx_fn}{path to Excel file}

\item{remove}{If TRUE, xlsx_fn is deleted upon execution.}
}
\description{
If the Excel is of length 1, it will be inputed into convert_xlsx_to_csv function. The tab names of
the Excel file are cleaned up using the rubix::cleanup_listnames function before being pasted into the final csv filename.
}
