% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek_cols.R
\name{peek_cols}
\alias{peek_cols}
\title{Retrieve Column Names in a Tabular File}
\usage{
peek_cols(file, tab = NULL, as_string = FALSE)
}
\description{
The source file may be an Excel or csv file.
}
\seealso{
\code{\link[stringr]{str_replace}}
}
