% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tribble.R
\name{print.tribble}
\alias{print.tribble}
\title{Print a tribble skeleton}
\usage{
\method{print}{tribble}(.data)
}
\arguments{
\item{.data}{data frame}
}
\description{
Print a tribble skeleton
}
\seealso{
\code{\link[dplyr]{mutate_all}}
}
