% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{read_cb_hl}
\alias{read_cb_hl}
\title{Read Headerless Data from Clipboard}
\usage{
read_cb_hl()
}
\value{
Dataframe is returned with all the rows and column names in the pattern "V{index}," where "{index}" is the position of the column.
}
\description{
Read headerless data from the clipboard. Shortcut for `read_clipboard(header = FALSE)`. To read data with a header from the clipboard, use \code{\link{read_cp}}.
}
\examples{
library(tidyverse)

# Test Data
test_data <-
        tibble(A = 1:3,
               B = c("Apples", "Bananas", "Oranges"),
               C = factor(c("Apples", "Bananas", "Oranges")),
               D = runif(3))


# Copy to Clipboard
copy_cb(data = test_data)

# Read Clipboard
read_cb()

# Check Datatypes
sapply(test_data, FUN = class)
sapply(read_cb(), FUN = class) # `read_cb()` function mutates all factors to character before returning

# Read Clipboard Shortcut for Headerless Data
read_cb_hl()

# Read Clipboard as a Tribble for Copy and Paste
read_cb_trbl()
read_cb_trbl(header = FALSE)
read_cb_trbl(quote = "'")
read_cb_trbl(header = FALSE, quote = "'")
}
\seealso{
Other clipboard functions: 
\code{\link{copy_cb}()},
\code{\link{copy_to_clipboard}()},
\code{\link{read_cb_trbl}()},
\code{\link{read_cb}()},
\code{\link{read_clipboard}()}
}
\concept{clipboard functions}
