% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_full_excel}
\alias{read_full_excel}
\title{Read an Entire Excel Workbook}
\usage{
read_full_excel(
  file,
  col_names = TRUE,
  col_types = "text",
  na = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl_progress(),
  .name_repair = "unique",
  log = "DEPRECATED",
  log_details = ""
)
}
\arguments{
\item{file}{File path.}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}

\item{col_types}{Either \code{NULL} to guess all from the spreadsheet or a
character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list". If exactly
one \code{col_type} is specified, it will be recycled. The content of a cell in
a skipped column is never read and that column will not appear in the data
frame output. A list cell loads a column as a list of length 1 vectors,
which are typed using the type guessing logic from \code{col_types = NULL}, but
on a cell-by-cell basis.}

\item{na}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}

\item{trim_ws}{Should leading and trailing whitespace be trimmed?}

\item{skip}{Minimum number of rows to skip before reading anything, be it
column names or data. Leading empty rows are automatically skipped, so this
is a lower bound. Ignored if \code{range} is given.}

\item{n_max}{Maximum number of data rows to read. Trailing empty rows are
automatically skipped, so this is an upper bound on the number of rows in
the returned tibble. Ignored if \code{range} is given.}

\item{guess_max}{Maximum number of data rows to use for guessing column
types.}

\item{progress}{Display a progress spinner? By default, the spinner appears
only in an interactive session, outside the context of knitting a document,
and when the call is likely to run for several seconds or more. See
\code{\link[readxl:readxl_progress]{readxl_progress()}} for more details.}

\item{.name_repair}{Handling of column names. By default, readxl ensures
column names are not empty and are unique. If the tibble package version is
recent enough, there is full support for \code{.name_repair} as documented in
\code{\link[tibble:tibble]{tibble::tibble()}}. If an older version of tibble is present, readxl falls
back to name repair in the style of tibble v1.4.2.}

\item{log}{(Deprecated) Log this function call to a log.md file?}

\item{log_details}{(Deprecated) If log is TRUE, additional notes to include with the log entry.}
}
\description{
Read an Entire Excel Workbook into a list of dataframes named by the original spreadsheet name. Arguments are passed to \code{\link[readxl]{read_excel}} for each sheet.
}
\seealso{
\code{\link[readxl]{excel_sheets}},\code{\link[readxl]{read_excel}}
 \code{\link[purrr]{set_names}},\code{\link[purrr]{map}}

Other excel functions: 
\code{\link{is_excel}()},
\code{\link{write_full_excel}()},
\code{\link{write_temp_xlsx}()}
}
\concept{excel functions}
\concept{read functions}
