% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-google.R
\name{read_gdoc_as_docx}
\alias{read_gdoc_as_docx}
\title{Read Google Doc as Officer Package Object}
\usage{
read_gdoc_as_docx(drive_path)
}
\arguments{
\item{drive_path}{Dribble, url, or id passed as `path` to the `googledrive package` functions`}
}
\description{
Read Google Doc as Officer Package Object
}
