% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{read_gdoc_as_txt}
\alias{read_gdoc_as_txt}
\title{Read Google Doc as Text}
\usage{
read_gdoc_as_txt(drive_path)
}
\arguments{
\item{drive_path}{Dribble, url, or id passed as `path` to the `googledrive package` functions`}
}
\description{
Read Google Doc as Text
}
