% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{simply_write_csv}
\alias{simply_write_csv}
\title{Write to CSV}
\usage{
simply_write_csv(
  x,
  file,
  na = "NA",
  append = FALSE,
  col_names = !append,
  quote_escape = "double",
  eol = "\\n",
  log = "DEPRECATED",
  log_details = ""
)
}
\arguments{
\item{log}{(Deprecated) Log this function call to a log.md file?}

\item{log_details}{(Deprecated) If log is TRUE, additional notes to include with the log entry.}
}
\description{
Wrapper to \code{\link[readr]{write_csv}}.
}
\seealso{
Other csv functions: 
\code{\link{is_csv}()},
\code{\link{simply_read_csv}()},
\code{\link{write_temp_csv}()}
}
\concept{csv functions}
