% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_as_xlsx}
\alias{view_as_xlsx}
\title{View in Temporary Excel File}
\usage{
view_as_xlsx(data)
}
\value{
Path to the temporary filename for unlinking
}
\description{
Write the data to a temporary xlsx file and open in Excel.
}
