% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_to_gfm.R
\name{word_to_gfm}
\alias{word_to_gfm}
\title{Converts all docx files in the working directory to markdown}
\usage{
word_to_gfm(docx)
}
\arguments{
\item{docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
