% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_to_md.R
\name{word_to_md}
\alias{word_to_md}
\title{Converts all docx files in the working directory to markdown}
\usage{
word_to_md(input_docx)
}
\arguments{
\item{docx}{name of the input docx filename}
}
\description{
Converts all docx files in the working directory to markdown
}
