% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx_if_not_exists.R
\name{write_xlsx_if_not_exist}
\alias{write_xlsx_if_not_exist}
\title{Writes xlsx to the OUTPUT directory with the option of having the activity logged and guardrails in case the chosen filename already exists}
\usage{
write_xlsx_if_not_exist(
  data,
  file,
  date_filename = FALSE,
  version = NULL,
  log = TRUE,
  io_comment = "",
  io_details = ""
)
}
\arguments{
\item{data}{dataframe or list object}

\item{date_filename}{if TRUE, appends filename with date and version number}

\item{log}{if TRUE, logs output in the log}

\item{file_basename}{basename of output file}
}
\description{
Writes xlsx to the OUTPUT directory with the option of having the activity logged and guardrails in case the chosen filename already exists
}
