% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_extension.R
\name{xlsx.writeMultipleData}
\alias{xlsx.writeMultipleData}
\title{FUNCTION_TITLE}
\usage{
xlsx.writeMultipleData(file, ...)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xlsx]{write.xlsx}}
}
