% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{Copy a data frame to the clipboard}
\usage{
copy_to_clipboard(data, log = "DEPRECATED", log_details = "")
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{log}{(Deprecated) Log this function call to a log.md file?}

\item{log_details}{(Deprecated) If log is TRUE, additional notes to include with the log entry.}
}
\description{
(Deprecated) Use \code{\link{copy_cb}} instead.
}
\seealso{
Other clipboard functions: 
\code{\link{copy_cb}()},
\code{\link{read_cb_hl}()},
\code{\link{read_cb_trbl}()},
\code{\link{read_cb}()},
\code{\link{read_clipboard}()}
}
\concept{clipboard functions}
