% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_full_excel}
\alias{read_full_excel}
\title{Reads entire XLSX file into a list of dataframes}
\usage{
read_full_excel(full_xlsx_fn, log = FALSE, log_details = "")
}
\arguments{
\item{full_xlsx_fn}{full path to xlsx filename}
}
\description{
Reads entire XLSX file into a list of dataframes
}
