% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simply_write_csv.R
\name{simply_write_csv}
\alias{simply_write_csv}
\title{A simple write to csv with a logging feature}
\usage{
simply_write_csv(x, path, ..., log_details = NULL)
}
\arguments{
\item{x}{data to write}

\item{path}{path to output file}
}
\description{
A simple write to csv with a logging feature
}
