% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_as_csv}
\alias{view_as_csv}
\title{View In Temporary CSV File}
\usage{
view_as_csv(data)
}
\value{
path to the temporary filename for unlinking
}
\description{
View a data in a csv opened in Excel without having to export to a permanant file.
}
