% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_as_csv.R
\name{view_as_csv}
\alias{view_as_csv}
\title{View dataframe in a temporary csv file}
\usage{
view_as_csv(dataframe)
}
\value{
path to the temporary filename for unlinking
}
\description{
View a dataframe in a csv opened in Excel without having to export to a permanant file.
}
