% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_as_xlsx.R
\name{view_as_xlsx}
\alias{view_as_xlsx}
\title{View dataframe in a temporary csv file}
\usage{
view_as_xlsx(dataframe)
}
\value{
path to the temporary filename for unlinking
}
\description{
View a dataframe in a csv opened in Excel without having to export to a permanant file.
}
