% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_as_xlsx}
\alias{view_as_xlsx}
\title{View a Dataframe in a Temporary Excel File}
\usage{
view_as_xlsx(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\value{
Path to the temporary filename for unlinking
}
\description{
View a dataframe in an Excel opened in Excel without having to export to a permanant file.
}
