% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_as_xlsx}
\alias{view_as_xlsx}
\title{View data in a temporary csv file}
\usage{
view_as_xlsx(data)
}
\value{
path to the temporary filename for unlinking
}
\description{
View a data in a csv opened in Excel without having to export to a permanant file.
}
