% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_full_excel.R
\name{write_full_excel}
\alias{write_full_excel}
\title{Write an Excel with a Log Entry}
\usage{
write_full_excel(x, file, log = TRUE, log_details = "", ...)
}
\arguments{
\item{x}{data to write, either as a list or a single dataframe}

\item{file}{path to xlsx file}

\item{log_details}{log detail or comment to include in the log entry. NULL if no log entry is required.}

\item{...}{optional parameters passed to openxlsx::write.xlsx}
}
\description{
Write an Excel with a Log Entry
}
