% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_temp_csv.R
\name{write_temp_csv}
\alias{write_temp_csv}
\title{A simply write csv to a temp file with a logging feature}
\usage{
write_temp_csv(x, ..., log_details = NULL)
}
\arguments{
\item{x}{data to write}
}
\description{
A call to simply_write_csv, returning the path to the temp file
}
