% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_temp_xlsx}
\alias{write_temp_xlsx}
\title{A simply write csv to a temp file with a logging feature}
\usage{
write_temp_xlsx(x, log = FALSE, log_details = "")
}
\arguments{
\item{x}{data to write}
}
\description{
A call to simply_write_csv, returning the path to the temp file
}
